requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
aniso8601==8.1.1
appdirs==1.4.4
appnope==0.1.0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
audioread==2.1.9
autopep8==1.5.4
awsebcli==3.19.4
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
biocommons.seqrepo==0.6.3.post1
bioutils==0.5.5
bleach==3.2.1
blessed==1.18.0
blis==0.4.1
bokeh==2.2.3
boto3==1.16.5
botocore==1.19.40
bs4==0.0.1
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.1.1
catalogue==1.0.0
cement==2.8.2
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
cld2-cffi==0.1.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
coloredlogs==15.0
configparser==5.0.2
contextvars==2.4
coverage==5.3
cryptography==3.2.1
cssselect==1.1.0
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
dask==2.30.0
dataclasses==0.7
ddt==1.2.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.2.9
distributed==2021.1.0
dnspython==2.1.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
email-validator==1.1.2
en-core-web-md==2.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.0.0/en_core_web_sm-2.0.0.tar.gz
entrypoints==0.3
es-core-news-md @ https://github.com/explosion/spacy-models/releases/download/es_core_news_md-2.0.0/es_core_news_md-2.0.0.tar.gz
et-xmlfile==1.0.1
fake-useragent==0.1.11
filelock==3.0.12
firebase-admin==4.3.0
Flask==1.0.2
Flask-Cache==0.13.1
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-WTF==0.14.3
fsspec==0.8.5
fudge==1.1.1
future==0.16.0
gensim==3.8.3
gevent==21.1.2
google-api-core==1.23.0
google-api-python-client==1.12.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.4.3
google-cloud-firestore==1.9.0
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
greenlet==1.0.0
grpcio==1.32.0
gspread==3.7.0
gspread-dataframe==3.1.1
gspread-formatting==0.3.7
HeapDict==1.0.1
helper==2.4.2
hgvs==1.5.1
httplib2==0.18.1
humanfriendly==9.1
idna==2.6
idna-ssl==1.1.0
ijson==3.1.4
imageio==2.9.0
immutables==0.14
importlib-metadata==2.1.1
imutils==0.5.4
install==1.3.4
ipykernel==5.3.4
ipyplot==1.1.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso4217==1.5.20180829
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-contrib-core==0.3.3
jupyter-core==4.6.3
jupyter-nbextensions-configurator==0.4.1
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
jupytext==1.6.0
kiwisolver==1.2.0
librosa==0.8.0
llvmlite==0.36.0
locket==0.2.0
locustio==0.9.0
lxml==4.6.2
Markdown==3.3.3
markdown-it-py==0.5.5
MarkupSafe==1.1.1
matplotlib==2.2.4
mistune==0.8.4
msgpack==0.6.2
msgpack-numpy==0.4.3.2
multidict==5.1.0
murmurhash==0.28.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.1
networkx==2.5
neuralcoref==4.0
nltk==3.5
notebook==6.1.4
numba==0.53.1
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.5.1.48
openpyxl==2.4.10
packaging==20.4
pandas==0.23.4
pandocfilters==1.4.2
paramiko==2.7.2
parse==1.19.0
Parsley==1.3
parso==0.7.1
partd==1.1.0
passlib==1.7.1
pathspec==0.5.9
pathy==0.3.6
patsy==0.5.1
pdfminer.six==20200517
pdfplumber==0.5.24
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
plac==0.9.6
plivo==4.15.2
pooch==1.3.0
praw==7.1.0
prawcore==1.5.0
preshed==1.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==3.13.0
psaw==0.0.12
psutil==5.8.0
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.9.9
pydantic==1.7.3
pydub==0.25.1
pyee==8.1.0
Pygments==2.7.1
PyJWT==1.7.1
pymusixmatch==0.3
PyNaCl==1.4.0
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyppeteer==0.2.5
pyquery==1.4.3
pyrsistent==0.17.3
pysam==0.16.0.1
PySocks==1.7.1
python-dateutil==2.8.1
python-Levenshtein==0.12.2
python-wordpress-xmlrpc==2.3
pytorch-lightning==1.1.4
pytube==10.5.1
pytube3==9.6.4
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
redis==3.4.1
regex==2017.4.5
requests==2.25.1
requests-html==0.10.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
s3transfer==0.3.3
sacremoses==0.0.43
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.3
seaborn==0.9.0
selenium==3.141.0
semantic-version==2.8.5
Send2Trash==1.5.0
sentencepiece==0.1.91
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
sklearn==0.0
smart-open==3.0.0
sortedcontainers==2.3.0
soundcloud==0.5.0
SoundFile==0.10.3.post1
soupsieve==2.1
spacy==2.0.12
spacy-legacy==3.0.1
sqlparse==0.4.1
srsly==1.0.5
statsmodels==0.12.0
stripe==2.20.3
symspellpy==6.7.0
tabulate==0.8.7
tblib==1.7.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
textblob==0.15.3
texttable==1.6.3
thinc==6.10.3
threadpoolctl==2.1.0
tifffile==2020.9.3
tika==1.24
tokenizers==0.10.2
toml==0.10.1
toolz==0.11.1
torch==1.7.0
torchvision==0.8.2
tornado==6.0.4
tqdm==4.50.2
traitlets==4.3.3
transformers==4.3.2
twilio==6.27.1
typer==0.3.2
typing-extensions==3.7.4.3
ua-parser==0.10.0
ujson==4.0.2
undetected-chromedriver==3.0.0
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.25.11
user-agents==1.1.0
w3lib==1.22.0
Wand==0.6.4
wasabi==0.8.2
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.10.11
WTForms==2.3.3
xgboost==1.2.1
xlrd==1.1.0
XlsxWriter==1.3.7
yarl==1.6.3
youtube-dl==2020.12.12
yoyo-migrations==7.3.2
zict==2.0.0
zipp==3.3.1
zope.event==4.5.0
zope.interface==5.2.0
